apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
xarray==2024.1.1
rioxarray==0.15.0
rasterio==1.3.10
shapely==2.0.4
geopandas==0.14.4
pyproj==3.6.1
pystac-client==0.7.6
planetary-computer==1.0.0
scikit-image==0.22.0
scipy==1.12.0
Pillow==10.3.0
Jinja2==3.1.4
rq==1.15.1
redis==5.0.4
pytest==8.1.1
